Switch to simpler base::WriteFile() variants (29/N)
This CL is part of a batch of CLs to replace the use of the less
friendly base::WriteFile call with simpler variants, which are easier to
read, and less prone to mistakes.

Changes in this particular CL address files under the path
/chrome/browser/profiles.

This CL was uploaded by git cl split.

R=skuhne@chromium.org

AX-Relnotes: n/a.
Bug: 418837
Change-Id: I45970bd803b12c51af5fbcc5067bdb5146eb52b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4285334
Auto-Submit: Claudio DeSouza <cdesouza@igalia.com>
Reviewed-by: Stefan Kuhne <skuhne@chromium.org>
Commit-Queue: Stefan Kuhne <skuhne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1110664}