Switch to simpler base::WriteFile() variants (109/N)
This CL is part of a batch of CLs to replace the use of the less
friendly base::WriteFile call with simpler variants, which are easier to
read, and less prone to mistakes.

Changes in this particular CL address files under the path
/extensions/common.

This CL was uploaded by git cl split.

R=rdevlin.cronin@chromium.org

AX-Relnotes: n/a.
Bug: 418837
Change-Id: Ie46a26307dbf5e5de6b65abc7d1952b84ce4d286
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4292077
Auto-Submit: Claudio DeSouza <cdesouza@igalia.com>
Commit-Queue: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1110717}