Remove logging::SetLogMessageHandler calls from RPH_WriteableFileTest
This removes a dependency on being able to cancel a LOG(FATAL) message
using a log-message handler. Removing these test dependencies are a
prerequisite for making LOG(FATAL) properly [[noreturn]].

This CL instead adds a testing callback for the handle checks, and uses
that in the test to verify the behavior.

BUG=1418101

Change-Id: Ica8ad45135263d0f5c080552f0ab3becf8acc889
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4292117
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/main@{#1110742}