[exo] Remove redundant version check
This CL removes the check that changed how output scale is being
calculated if the interface is bound at a version < 6.

This was added back in 2018 and was meant to be removed in a couple
of weeks after the client version had caught up. See comment below
for context.

https://chromium-review.googlesource.com/c/chromium/src/+/1049115/comment/22f3e466_f6086586/

Bug: None
Change-Id: I1aa38a92d40d1759f3bfd260cd63a9f7bdd19a05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4284823
Auto-Submit: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1110758}