Switch to simpler base::WriteFile() variants (64/N)
This CL is part of a batch of CLs to replace the use of the less
friendly base::WriteFile call with simpler variants, which are easier to
read, and less prone to mistakes.

Changes in this particular CL address files under the path
/chromeos.

This CL was uploaded by git cl split.

R=hashimoto@chromium.org, wenbojie@chromium.org

AX-Relnotes: n/a.
Bug: 418837
Change-Id: I5bbb9ab442f44fe1783421b78a2c91663092dbe0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4290385
Commit-Queue: Ryo Hashimoto <hashimoto@chromium.org>
Auto-Submit: Claudio DeSouza <cdesouza@igalia.com>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1110769}