Set up synthetic field trial for hugepage zygote remap.
The hugepage zygote remap feature was historically gated behind the
field trial CrOSHugepageRemapAndLockInZygote. This logic has never
worked (and will never work) due to the timing of zygote Finch
initialization. This CL removes the Finch feature and instead sets up a
synthetic field trial to do a 50/50 experiment on the feature.

Historically the feature was default-on in tip of tree. We tried setting
the feature to default-off in tip of tree for ~3 weeks and there were no
significant functional or performance regressions. So having tip of tree
reflect a non-deterministic state (sometimes off, sometimes on) should
be safe.

Bug: 1402565
Test: Manually load a build with custom logging onto a ChromeOS device.
Change-Id: I809fbb509f5c3c354ce32657fbf4254eb41c628a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4269691
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/main@{#1110785}