exo: Borealis: make aura_toplevel.set_window_bounds() set client bounds.
This allows set_window_bounds() to be safely used for decorated
windows. The change is restricted to Borealis per request.

Core Wayland protocols ignore window decorations entirely; clients
deal only with surface bounds, and decorations are left to the
compositor. If we want Exo to permit clients to set the window
bounds, we can either:

(1) work around this by building a custom protocol to tell clients
    the size of the decorations (similar to _NET_FRAME_EXTENTS in X11)
    and change configure() events to send window bounds instead of
    client bounds; or
(2) make Exo expand the requested bounds to account for the
    decorations itself.

This CL chooses option 2.

This does mean that clients may position the window such that the
frame is offscreen, and indeed this happened during my testing.
To prevent this, push windows down to ensure their top is at least
equal to the top of their output.

BUG=b:247452928
TEST=Patch the set_window_bounds version of https://crrev.com/c/3900105 and run Chrome with DCHECKs enabled. Observe: (1) Chrome no longer DCHECKs when running Borealis apps; (2) Path of Exile is not offset to the bottom-right in windowed mode; (3) Stardew Valley maintains a consistent size in windowed mode, rather than shrinking vertically due to set_window_bounds() and configure() disagreeing about whether to include window bounds.

Change-Id: Ie3acfb2597fad5f6da4bb4d54f43b3fff33529cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4234678
Commit-Queue: Chloe Pelling <cpelling@google.com>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1110791}