[white-space] Disables a use of `white-space` in `editing_style`
This patch disables a piece of code in the `StyleChange`
constructor in `editing_style.cc`.

As part of making the `white-space` property to a shorthand,
this code needs a change, but the desired change isn't obvious
because no tests fail without this.

From the git log, this was added in 2006[1] with the
`pasting-tabs.html` test. Then it was refactored in 2009[2]
and then moved from `ApplyStyleCommand` in 2011[3].

In ToT, the `kWhiteSpace` code in
`ConflictsWithInlineStyleOfElement`[4] is relevant to pass
this test. It looks like this code is a left over.

If regressions are seen, a runtime flag is added to enable it
back, and we will know how to rewrite this code when the
`white-space` becomes a shorthand.

[1] https://chromium.googlesource.com/chromium/src/+/30d7436a45466666b14f85efb3ab1413682b1fe9%5E%21
[2] https://chromium.googlesource.com/chromium/src/+/75d446bf55f5306e3c7bbb7ff620dd0142ad8530%5E%21/
[3] https://bugs.webkit.org/attachment.cgi?id=85098&action=prettypatch
[4] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/editing/editing_style.cc;l=1007?q=EditingStyle::ConflictsWithInlineStyleOfElement&ss=chromium

Bug: 1417543
Change-Id: Icd34143ba14e68cf138591ea3068d8338d091f56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4289333
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1110798}