Compile manual-paths-to-ignore.txt into the clang plugin
Specifying the list of paths to exclude using a text file didn't work
with goma because the file wouldn't be on the server.

This CL compiles in the list as part of the clang plugin.
raw-ptr-exclude-path= arg can be used to append to the compiled in list
while waiting for the clang roll.

Change-Id: Ia63884a6c456b35ba0c9c3fe96248312352068b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4248075
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Keishi Hattori <keishi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1110821}