[bfcache] Clean up `kBackForwardCache` feature's usage in tests.
Due to historical reason, the feature flag for BackForwardCache, i.e.
`kBackForwardCache`, has grown into a complex feature that supports
multiple parameters. To enable the feature in a test and make it stable,
one is required to set up extra parameters, and also required to disable
another feature `kBackForwardCacheMemoryControls`. The whole setup is so
complicated that some tests made wrong usage of them.

This CL is the 1st step to reduce its complexity, by hiding the setup behind team-owned utility functions. In the future, the team may decide to refactor the parameters, and this change help limiting the affected files.

The changes include:
- Rename test utility functions to make them easier to understand.
- Replace most references to `kBackForwardCache` or `kBackForwardCacheMemoryControls` in tests with the above functions

Bug: 1301867
Change-Id: I117196bf639c8e88adc0d813b5f225040516fd0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4280121
Reviewed-by: Ken Okada <kenoss@chromium.org>
Reviewed-by: Matthias Körber <koerber@google.com>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Ming-Ying Chung <mych@chromium.org>
Reviewed-by: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1110890}