Reset threshold for accepting Autofill popup entries on view update.
Currently, the Autofill controller measures when a popup view is shown
the first time and measures against this point in time when checking
whether the necessary time threshold is met in "AcceptSuggestion".
This CL makes the change to update the time measurement also if the
view is updated. In this way, the threshold cannot be circumvented
by showing the popup in one part of the screen and then clicking
somewhere else so that the popup's coordinates are updated.

Bug: 1411172, 1418837
Change-Id: Ice0acc8d11fddf0d8d32a1a22b40e15c6941285a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4295697
Commit-Queue: Jan Keitel <jkeitel@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1110907}