Don't clear entropy state prefs for FRE metrics opt-out
If the user opted out of collecting metrics in the First Run
Experience - do not clear entropy state prefs. This is safe to do,
since no data has been uploaded yet.

This should fix the issue when user groups are reassigned on the second
run if the user opts out of UMA.

Bug: 1416564
Change-Id: I7fb28390f53018b652402099b662f3b8627a93f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4255472
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1110948}