Stub out flaky NetworkConnectionTrackerBrowserTest on MacOS and CrOS
Stub out NetworkConnectionTrackerBrowserTest.SimulateNetworkServiceCrash
on MacOS and CrOS to match the test for NetworkConnectionTracker.

If the connection tracker itself shouldn't be tested on MacOS because
of browser test fixture process issues on MacOS then the deeper test
of crashing the network service to observe the impact on connection
tracker should also not be tested there.

This likely explains why the test has been flaky on MacOS for at least
the last 90 days but likely much longer (seems to have gotten worse
recently).

Bug: 1419812
Change-Id: Icb9427c2792e501d35754e3eb7a02ea63e2a72a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4294460
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: Patrick Meenan <pmeenan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1110954}