Fix TouchToFillRenderTest.testShowsOneCredentialModernHalfState
The test testShowsOneCredentialModernHalfState was disabled in
crbug.com/1416856 because it was suspected it's still flaky,
even after the fixing of the loading bar.
The problem was not the loading bar, but the flakiness of the URL
that was displayed for the credential. The URL sometimes contained "https" and sometimes not and that was the topic of crbug.com/1414577.
The URL actually should not have been on the credential items of that
test because it should be shown for PSL matches and affiliated credentials only.
The URL was most likely flaky because it wasn't formatted with
UrlFormatter.formatUrlForDisplayOmitScheme like it is in the production
code.
The URL on credential items is removed from the currently existing tests
and the case of affiliated credentials will be tested separately.

Old screenshot: https://screenshot.googleplex.com/A9wJNac8DQZHPTA
New screenshot: https://screenshot.googleplex.com/5WHYJodD39FE3QT

Bug: 1414577, 1416856
Change-Id: I648e10d5529fbae63c4bc264a8d96a2a0ce6d8d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4295657
Reviewed-by: Friedrich Horschig <fhorschig@chromium.org>
Commit-Queue: Ivana Žužić <izuzic@google.com>
Cr-Commit-Position: refs/heads/main@{#1110990}