Switch to simpler base::WriteFile() variants (129/N)
This CL is part of a batch of CLs to replace the use of the less
friendly base::WriteFile call with simpler variants, which are easier to
read, and less prone to mistakes.

Changes in this particular CL address files under the path
/services/network.

This CL was uploaded by git cl split.

R=ericorth@chromium.org

AX-Relnotes: n/a.
Bug: 418837
Change-Id: Ia4cb106c284548a48e7bd75cec149005391015a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4292133
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Auto-Submit: Claudio DeSouza <cdesouza@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1110991}