Switch to simpler base::WriteFile() variants (20/N)
This CL is part of a batch of CLs to replace the use of the less
friendly base::WriteFile call with simpler variants, which are easier to
read, and less prone to mistakes.

Changes in this particular CL address files under the path
/chrome/browser/net.

This CL was uploaded by git cl split.

R=ericorth@chromium.org

AX-Relnotes: n/a.
Bug: 418837
Change-Id: I3ab21596506cb81d9cb3eab9f8f341ada4db7b02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4285487
Auto-Submit: Claudio DeSouza <cdesouza@igalia.com>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1110995}