[QuickStart] Added QuickStartDecoder to Nearby Service
This change adds QuickStartDecoder to the Nearby Service, allowing it to
be initialized and used.

QuickStartDecoder was set up as a Mojo service, but didn't have the proper
wiring and placement to be launched. This change makes the required file
moves and changes in order to make it launch-able

Change-Id: I4cf4111e2a55e0d5e38fa05305b7d24ea2200fc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4277302
Reviewed-by: Michael Hansen <hansenmichael@google.com>
Reviewed-by: Brittany Hartmire <bhartmire@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Joe Antonetti <joeantonetti@google.com>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1111022}