Switch to simpler base::WriteFile() variants (62/N)
This CL is part of a batch of CLs to replace the use of the less
friendly base::WriteFile call with simpler variants, which are easier to
read, and less prone to mistakes.

Changes in this particular CL address files under the path
/chromecast/crash.

This CL was uploaded by git cl split.

R=dnicoara@chromium.org

AX-Relnotes: n/a.
Bug: 418837
Change-Id: I916b1ada29bc2e54f2f39e6177c63f6bb15c2077
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4288686
Auto-Submit: Claudio DeSouza <cdesouza@igalia.com>
Commit-Queue: Daniel Nicoara <dnicoara@chromium.org>
Reviewed-by: Daniel Nicoara <dnicoara@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1111031}