Lacros: Restore all previous profile for if profile picker is disabled.
Previously on Lacros if profile picker is disabled and the browser is
crashed with multiple profiles windows open, only the main profile
window will open and show the recover bubble. Other profiles are not
able to recover from the previous crash. This CL routes
NewWindowWithProfile() and LaunchOrNewTabWithProfile() to call the full
restore API when there is a crash detected so one window for each
profile can be open and all profiles can be recovered from the crash.

TEST=out_linux_lacros/Release/bin/run_lacros_chrome_browsertests_run_in_series --gtest_filter="*BrowserLauncherTest.OpenTwoWindowsAfterCrashWithTwoProfiles*"

Bug: 1409325
Change-Id: If19f11ea88881e74e9ea6d51da6f9e8ddf25648c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4245827
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Yuheng Huang <yuhengh@chromium.org>
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1111065}