Switch to simpler base::WriteFile() variants (30/N)
This CL is part of a batch of CLs to replace the use of the less
friendly base::WriteFile call with simpler variants, which are easier to
read, and less prone to mistakes.

Changes in this particular CL address files under the path
/chrome/browser/resource_coordinator.

This CL was uploaded by git cl split.

R=chrisha@chromium.org

AX-Relnotes: n/a.
Bug: 418837
Change-Id: Ibf08193be48f77acb9c4b6ef168aca78a84efe1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4285335
Commit-Queue: David Roger <droger@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Auto-Submit: Claudio DeSouza <cdesouza@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1111068}