Hook dropping animation for AppsGridView with DragDropController
When an app item is dropped into the apps grid, it is expected that the
drag image animates smoothly into the final bounds.  Recent effort are
trying to implement the drag and drop API into the app list drag and
drop flow. In result, the drag image proxy will be removed in favor of
the DragImage from the Drag and Drop controller.  To successfully
maintain the same animation, the apps grid view needs to be able to
manipulate the drag image widget only for this case.

CL http://crrev.com/c/4264073 introduces GetDropCallbackWithAnimation
method to fetch the  drag image widget from the drag controller at drop
point. This CL implements said method to hook the existing dropping
animation on the AppsGridView.

Bug: b:261985897
Change-Id: I20aa0d50396c4223cdabe310703e332b17c98207
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4228794
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1111111}