[Notification Drag] Implement image notification drag - Part 7
This CL removes the dragged notification by calling MessageCenter::
RemoveNotification() if the notification drag-and-drop ends with
data copy.

Note that drop could be async, which means that a new drag-and-drop
session could start before the active async drop completes. Therefore,
in this CL, AshNotificationDragController cleans up the data members
for the active drag-and-drop session when:

1. Drag is cancelled; or
2. Drop is completed; or
3. A new drag-and-drop session starts without waiting for the async
drop of the previous drag-and-drop session to finish. DragDrop-
Controller guarantees that drag-and-drop observers are not notified of
async drop that is interrupted by a new drag-and-drop.

This CL adds unit tests to cover notification sync/async drops.

Bug: b/264934296
Change-Id: I7a77a10e4e06a51afc56641f4512118519be82a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4226622
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Andrew Xu <andrewxu@google.com>
Reviewed-by: Ahmed Mehfooz <amehfooz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1111130}