Switch to simpler base::WriteFile() variants (10/N)
This CL is part of a batch of CLs to replace the use of the less
friendly base::WriteFile call with simpler variants, which are easier to
read, and less prone to mistakes.

Changes in this particular CL address files under the path
/chrome/browser/enterprise.

This CL was uploaded by git cl split.

R=seblalancette@chromium.org

AX-Relnotes: n/a.
Bug: 418837
Change-Id: I6d95b683b0304efc0a954f607fd9452fa6a7fe58
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4285245
Auto-Submit: Claudio DeSouza <cdesouza@igalia.com>
Reviewed-by: Roger Tawa <rogerta@chromium.org>
Commit-Queue: Claudio DeSouza <cdesouza@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1111149}