[Auto-disable Accessibility] Improve WebContentsAccessibilityImpl obscuring by client views
This CL updates the implementation of WebContentsAccessibility to
have a stricter use of mIsObscuredByAnotherView.

With this CL we update the implementation of obscuring the web
contents for accessibility to include an assert that only one
client is modifying this variable at a time. The behavior when
two views try to obscure content, and one dismisses, is not well
defined. In such a case we should throw an AssertionError and
handle it on a case-by-case basis.

With this CL we make some updates to clients to ensure that they
are not both modifying the value simultaneously. We also update
the impls to properly handle maintaining proper state during
web contents switches.

When the WCAI is obscured by another view, i.e.
mIsObscuredByAnotherView=true, then WCAI will return a null
AccessibilityNodeProvider and not respond to performAction calls
sent by the framework.

WCAI no longer observe changes in accessibility state directly, it
relies on listening to the AccessibilityState class, which is the
source-of-truth going forward for a11y state info.

This CL should cause no functional changes.


AX-Relnotes: N/A
Bug: 1362100
Change-Id: Id077b7f5fb1d587e408ed5635ad33f7e78128aac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4205857
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Amanda Lin Dietz <aldietz@google.com>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1111156}