Switch to simpler base::WriteFile() variants (76/N)
This CL is part of a batch of CLs to replace the use of the less
friendly base::WriteFile call with simpler variants, which are easier to
read, and less prone to mistakes.

Changes in this particular CL address files under the path
/components/keyed_service.

This CL was uploaded by git cl split.

R=hnakashima@chromium.org

AX-Relnotes: n/a.
Bug: 418837
Change-Id: Ie740dbe50b8e940a4132caaeed56b3e39414e11a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4291861
Auto-Submit: Claudio DeSouza <cdesouza@igalia.com>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1111158}