Flaky Test: TabGridIphTest
A number of tests in this suite are disabled. Some failures look to be
possibly related to not waiting long enough for the view before acting
on it. Change transitioning to the GTS to use onViewWaiting.

Don't enable any tests until test-reviver says it is safe to do so.
I'll follow-up in a week to see how the tests are performing on the
reviver.

Bug: 1412947, 1412394, 1300743, 1245260, 1381298
Change-Id: I6b19858223a7fea5dd31e4f1d4a5a22d475a8a3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4295665
Auto-Submit: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Xi Han <hanxi@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Xi Han <hanxi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1111173}