Re-introduce constructors to InterestGroup.
A previous CL[1] removed the constructors for InterestGroup to allow for
named aggregation initialization[2]. However, that is a C++20 feature,
and is disallowed in the Chromium codebase. This CL removes that and
reverts the InterestGroup constructor calls to what they originally
were.

[1]https://chromium-review.googlesource.com/c/chromium/src/+/4083309
[2]https://chromium-review.googlesource.com/c/chromium/src/+/4083309/comment/54698eab_5c4245d5/

Change-Id: I6d7b6d7acee8dd46ad44b89a6289ff98ad0276ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4294485
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Liam Brady <lbrady@google.com>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1111184}