Switch to simpler base::WriteFile() variants (45/N)
This CL is part of a batch of CLs to replace the use of the less
friendly base::WriteFile call with simpler variants, which are easier to
read, and less prone to mistakes.

Changes in this particular CL address files under the path
/ash/projector.

This CL was uploaded by git cl split.

R=xiqiruan@chromium.org

AX-Relnotes: n/a.
Bug: 418837
Change-Id: I6e14b212e81d53dc24d7e84bf9ec4e58afe1c694
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4291973
Reviewed-by: Li Lin <llin@chromium.org>
Auto-Submit: Claudio DeSouza <cdesouza@igalia.com>
Commit-Queue: Li Lin <llin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1111205}