Fix WebView Fullscreen Mismatched Orientation Transition
WebView can be embedded in a Fullscreen application, without itself
being Fullscreen.

If the WebView is of a different Orientation from that of the
application, it can end up failing to synchronize. The handling of
mismatched orientations of the PhysicalBackingSize and ScreenInfo was
missing removing a throttle used for initial fullscreen transitions.

This change re-adds it. This change also fixes the parameterization in
the tests, which didn't have the Disable path updated when the flag
was changed to be on by default

TEST=*RenderWidgetHostViewAndroid*

Bug: 1418321
Change-Id: I2cfb0f404a3f8a4a958cebf9853f1ed2044cb18c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4297803
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1111225}