[DownloadBubble] Load file type icon unconditionally when setting icon
This fixes an issue where rows with non-filetype icons (i.e. rows with
an incognito icon or warning icon) did not cause the IconManager to
load the filetype icon. This meant that the filetype icon did not get
cached in the IconManager, causing a subsequent lookup before starting
drag-and-drop to return null.

This CL fixes that by always looking up the filetype icon, and storing
it as a member of DownloadBubbleRowView for later use as the drag icon.

This also fixes the fact that, if there is no filetype icon, we would
also always get nullptr from the lookup before drag-and-drop, which is
no good. Now, in the case where there is no filetype icon, we use a
default icon.

Also retires a histogram that is not actionable, and is no longer
convenient to log only once per SetIcon().

Bug: 1417326
Change-Id: If1ef875faaa30a6840bab19a973ec5869fc7060a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4289231
Commit-Queue: Lily Chen <chlily@chromium.org>
Reviewed-by: Min Qin <qinmin@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1111233}